Switch bootsplashes for pinephone and pinetab
This commit adds a separate splash for the PinePhone and PineTab as SVG
files. The build-initrd.sh script then converts these files into 32-bit
RGBA files for use in the initrd.

This completely breaks building the initramfs from a running system,
it can now only be done via the builder script.