[Do not merge] Resize gpt when needed
The PinePhone and PineTab shipped with eMMC larger than the image that
shipped on them. The PineTab in particular shipped with a 64GB eMMC and
a 16GB image. This change adds a quick

This is inappropriate for merge as it assumes the system disk is
/dev/mmcblk(something), requires use of the devnum environment variable,
and assumes that userdata would be partition 10 on the media.